+++ src/util.ts
@@ -4,3 +4,21 @@ export const useForceUpdate = () => {
   const [, updateState] = useState<any>({})
   return useCallback(() => updateState({}), [])
 }
+
+// Promise wrapper
+
+export class Deferred<T> {
+  promise: Promise<T> = new Promise((resolve, reject) => {
+    this.resolve = value => {
+      this.settled = true
+      resolve(value)
+    }
+    this.reject = reason => {
+      this.settled = true
+      reject(reason)
+    }
+  })
+  settled = false
+  resolve: (value: T | PromiseLike<T>) => void
+  reject: (error: Error) => void
+}